MultiQueue | fixes

+++ src/services/subscription/subscriptions.rs
@@ -1,7 +1,7 @@
-use std::{fmt::Debug, hash::BuildHasherDefault, sync::Arc};
+use std::{fmt::Debug, hash::BuildHasherDefault, sync::{Arc, atomic::{AtomicUsize, Ordering}}};
 use hashers::fx_hash::FxHasher;
 use sal_core::error::Error;
-use crate::{collections::FxDashMap, services::entity::Point, sync::{RwLock, channel::Sender}};
+use crate::{collections::{FxDashMap, FxDashSet}, services::entity::Point, sync::{RwLock, channel::Sender}};
 ///
 /// Unique id of the service (TxId) receiving the Point's by the subscription
 /// This id used to identify the service produced the Points. 
@@ -20,11 +20,10 @@ type SubscriberList = Arc<Vec<(ReceiverId, Sender<Point>)>>;
 ///
 /// Contains map of Sender's
 /// - Where Sender - is pair of String ID & Sender<PointType>
-#[derive(Clone)]
 pub struct Subscriptions {
     dbg: String,
     /// Справочник для расширения Multicast подписок (добавление новых PointDest к существующим получателям)
-    registry: FxDashMap<ReceiverId, Sender<Point>>,
+    registry: FxDashMap<ReceiverId, Arc<ReceiverInfo>>,
     /// Multicast подписки (по конкретным PointDest)
     multicast: FxDashMap<PointDest, SubscriberList>,
     /// Broadcast подписки (На все возможные PointDest)
@@ -48,11 +47,10 @@ impl Subscriptions {
     /// 
     /// COW Write: Медленнее (Copy), но безопасно для Readers
     pub fn add_multicast(&self, receiver_id: ReceiverId, destination: &str, sender: Sender<Point>) {
-        // 1. Обновляем Registry новым получателем для будущего возможного расширения подписки
-        self.registry.entry(receiver_id).or_insert(sender.clone());
-        // 2. Атомарная транзакция для Multicast
+        // Атомарная транзакция для Multicast
         // DashMap держит Write Lock на этот бакет, пока выполняется замыкание.
         // Никто другой не сможет прочитать или записать в этот ключ, пока мы не закончим.
+        let mut added = false;
         self.multicast.entry(destination.to_owned())
             .and_modify(|arc_vec| {
                 // Внутри этого блока мы в безопасности (Critical Section)
@@ -63,12 +61,20 @@ impl Subscriptions {
                     new_vec.push((receiver_id, sender.clone()));
                     // 3. Подменяем обновленный массив получателей
                     *arc_vec = Arc::new(new_vec);
+                    added = true;
                 }
             })
             .or_insert_with(|| {
+                added = true;
                 // Если ключа не было - создаем новый
-                Arc::new(vec![(receiver_id, sender)])
+                Arc::new(vec![(receiver_id, sender.clone())])
             });
+        if added {
+            // Обновляем Registry новым получателем для будущего возможного расширения подписки
+            self.registry.entry(receiver_id)
+                .and_modify(|receiver_info| receiver_info.inc(Some(destination)))
+                .or_insert(Arc::new(ReceiverInfo::new(sender, Some(destination))));
+        }
     }
     ///
     /// Extends subscription for `receiver_id` if exists, otherwise returns error
@@ -76,8 +82,9 @@ impl Subscriptions {
         let error = Error::new(&self.dbg, "extend_multicast");
         log::trace!("{}.extend_multicast | Extending (multicast) for receiver: {} ({})...", self.dbg, destination, receiver_id);
         // 1. Берем из Registry получателя если такой есть
-        match self.registry.get(&receiver_id).map(|s| s.clone()) {
-            Some(sender) => {
+        let mut extended = false;
+        match self.registry.get(&receiver_id).map(|receiver_info| receiver_info.clone()) {
+            Some(receiver_info) => {
                 // 2. Атомарная транзакция для Multicast
                 self.multicast.entry(destination.to_owned())
                     .and_modify(|arc_vec| {
@@ -86,15 +93,20 @@ impl Subscriptions {
                         if !arc_vec.iter().any(|(id, _)| *id == receiver_id) {
                             // 2. Клонируем массив получателей если нужно добавить
                             let mut new_vec = (**arc_vec).clone();
-                            new_vec.push((receiver_id, sender.clone()));
+                            new_vec.push((receiver_id, receiver_info.sender()));
                             // 3. Подменяем обновленный массив получателей
                             *arc_vec = Arc::new(new_vec);
+                            extended = true;
                         }
                     })
                     .or_insert_with(|| {
                         // Если ключа не было - создаем новый
-                        Arc::new(vec![(receiver_id, sender)])
+                        extended = true;
+                        Arc::new(vec![(receiver_id, receiver_info.sender())])
                     });
+                if extended {
+                    receiver_info.inc(Some(destination));
+                }
                 log::trace!("{}.extend_multicast | Extending (multicast) for receiver: {} ({}) - Ok", self.dbg, destination, receiver_id);
                 Ok(())
             }
@@ -110,10 +122,14 @@ impl Subscriptions {
         let mut lock = self.broadcast.write();
         let mut new_vec = (**lock).clone();
         match new_vec.iter_mut().find(|(id, _)| *id == receiver_id) {
-            Some((_, old_sender)) => *old_sender = sender,
-            None => new_vec.push((receiver_id, sender)),
+            Some((_, old_sender)) => *old_sender = sender.clone(),
+            None => new_vec.push((receiver_id, sender.clone())),
         }
         *lock = Arc::new(new_vec);
+        // Обновляем Registry новым получателем для будущего возможного расширения подписки
+        self.registry.entry(receiver_id)
+            .and_modify(|receiver_info| receiver_info.inc(None::<PointDest>))
+            .or_insert(Arc::new(ReceiverInfo::new(sender, None::<PointDest>)));
     }
     ///
     /// ## Returns all pairs of `ReceiverId`, `Sender`'s for the specified `point_id`
@@ -151,66 +167,40 @@ impl Subscriptions {
     }
     ///
     /// Removes single subscription by Point Id for receiver ID
-    pub fn remove(&self, receiver_id: ReceiverId, destinations: impl IntoIterator<Item = impl Into<PointDest>>) -> Result<(), Error> {
-        let error = Error::new(&self.dbg, "remove");
-        let mut errors = vec![];
-        let mut removed = false;
+    pub fn remove(&self, receiver_id: ReceiverId, destinations: impl IntoIterator<Item = impl Into<PointDest>>) {
         for destination in destinations {
             let destination: PointDest = destination.into();
             // Атомарная транзакция для Multicast
             // DashMap держит Write Lock на этот бакет, пока выполняется замыкание.
             // Никто другой не сможет прочитать или записать в этот ключ, пока мы не закончим.
-            match self.multicast.entry(destination.clone()) {
-                dashmap::Entry::Occupied(mut entry) => {
-                    let arc_vec = entry.get_mut();
-                    // .and_modify(|arc_vec| {
-                        // Внутри этого блока мы в безопасности (Critical Section)
-                        // Ищем получателя которого надо удалить
-                    match arc_vec.iter().position(|(id, _)| *id == receiver_id) {
-                        Some(pos) => {
-                            // Клонируем массив получателей если нужно удалить
-                            let mut new_vec = (**arc_vec).clone();
-                            new_vec.remove(pos);
-                            match new_vec.is_empty() {
-                                // Подменяем обновленный массив получателей
-                                false => *arc_vec = Arc::new(new_vec),
-                                // Удаляем массив получателей если он пуст
-                                true => _ = entry.remove_entry(),
-                            }
-                            removed = true;
-                        }
-                        _ => errors.push(format!("Subscription '{destination}' - NOT FOUND for receiver '{receiver_id}'")),
+            if let dashmap::Entry::Occupied(mut entry) = self.multicast.entry(destination.clone()) {
+                let arc_vec = entry.get_mut();
+                // .and_modify(|arc_vec| {
+                    // Внутри этого блока мы в безопасности (Critical Section)
+                    // Ищем получателя которого надо удалить
+                if let Some(pos) = arc_vec.iter().position(|(id, _)| *id == receiver_id) {
+                    // Клонируем массив получателей если нужно удалить
+                    let mut new_vec = (**arc_vec).clone();
+                    new_vec.remove(pos);
+                    match new_vec.is_empty() {
+                        // Подменяем обновленный массив получателей
+                        false => *arc_vec = Arc::new(new_vec),
+                        // Удаляем массив получателей если он пуст
+                        true => _ = entry.remove_entry(),
+                    }
+                    // Удаляем из Registry получателя
+                    if let Some(receiver_info) = self.registry.get(&receiver_id) {
+                        receiver_info.dec(Some(destination))
                     }
                 }
-                _ => errors.push(format!("Subscription '{destination}' - NOT FOUND")),
             }
         }
-        if removed && !self.is_subscribed(receiver_id) {
-            // Удаляем из Registry если получателя больше нет в подписках
-            self.registry.remove(&receiver_id);
-        }
-        match removed {
-            true => Ok(()),
-            false => Err(error.err(errors.join("\n"))),
-        }
+        // Удаляем из Registry если получателя больше нет в подписках
+        self.registry.remove_if(&receiver_id, |_, r| !r.is_subscribed());
     }
     ///
     /// Removes all subscriptions for `receiver_id`
-    pub fn remove_all(&self, receiver_id: ReceiverId) -> Result<(), Error> {
-        let error = Error::new(&self.dbg, "remove_all");
-        let mut err = None;
-        let destinations: Vec<PointDest> = self.multicast.iter()
-            .filter(|entry| entry.value().iter().any(|(id, _)| *id == receiver_id))
-            .map(|entry| entry.key().clone())
-            .collect();
-        if !destinations.is_empty() {
-            if let Err(e) = self.remove(receiver_id, &destinations) {
-                err = Some(error.pass(e));
-            }
-        }
-        if !self.is_subscribed(receiver_id) {
-            self.registry.remove(&receiver_id);
-        }
+    pub fn remove_all(&self, receiver_id: ReceiverId) {
         // Удаляем Broadcast подписку для получателя (receiver_id)
         let mut lock= self.broadcast.write();
         if let Some(pos) = lock.iter().position(|(id, _)| *id == receiver_id) {
@@ -219,19 +209,30 @@ impl Subscriptions {
             new_vec.remove(pos);
             // 3. Подменяем обновленный массив получателей
             *lock = Arc::new(new_vec);
+            self.registry.entry(receiver_id)
+            .and_modify(|receiver_info| receiver_info.dec(None::<PointDest>));
         }
-        match err {
-            None => Ok(()),
-            Some(err) => Err(err),
+        let destinations: Vec<PointDest> = self.registry.get(&receiver_id)
+            // .filter(|entry| entry.value().iter().any(|(id, _)| *id == receiver_id))
+            .map(|entry| entry.destinations())
+            .unwrap_or_default();
+        if !destinations.is_empty() {
+            self.remove(receiver_id, &destinations);
+        } else {
+            self.registry.remove_if(&receiver_id, |_, r| !r.is_subscribed());
         }
     }
     ///
     /// Returns true if [ReceiverId] has any subscriptions
     pub fn is_subscribed(&self, receiver_id: ReceiverId) -> bool {
-        if self.broadcast.read().iter().any(|(id, _)| *id == receiver_id) {
-            return true;
-        }
-        self.multicast.iter().any(|entry| entry.iter().any(|(id, _)| *id == receiver_id))
+        self.registry
+            .get(&receiver_id)
+            .map(|r| r.is_subscribed())
+            .unwrap_or(false)
+        // if self.broadcast.read().iter().any(|(id, _)| *id == receiver_id) {
+        //     return true;
+        // }
+        // self.multicast.iter().any(|entry| entry.iter().any(|(id, _)| *id == receiver_id))
     }
     ///
     /// Removes all subscriptions
@@ -266,5 +267,66 @@ impl SubscribersView {
         // Создаем цепочку итераторов: сначала broadcast, потом multicast (если есть)
         let mc_iter = self.multicast.as_deref().into_iter().flat_map(|v| v.iter());
         self.broadcast.iter().chain(mc_iter)
+    }
+}
+///
+/// Контейнер для хранения информации о подписчике
+#[derive(Debug)]
+struct ReceiverInfo {
+    /// Sender получателя, `subscriptions` раз упомянутый в Broadcast и Multicast подписках
+    sender: Sender<Point>,
+    /// Количество подписок (Broadcast + Multicast)
+    subscriptions: AtomicUsize,
+    /// Нименования всех подписок получателя, что бы не искать их при удалении
+    destinations: FxDashSet<PointDest>,
+}
+//
+impl ReceiverInfo {
+    ///
+    /// Returns [ReceiverInfo] new instance
+    /// - `sender` - Channel to the coresponding Receiver
+    pub fn new(sender: Sender<Point>, destination: Option<impl Into<PointDest>>) -> Self {
+        let destinations = FxDashSet::default();
+        if let Some(dest) = destination {
+            destinations.insert(dest.into());
+        }
+        Self {
+            sender,
+            subscriptions: AtomicUsize::new(1),
+            destinations,
+        }
+    }
+    ///
+    /// Returns clone of the Receiver's Sender
+    pub fn sender(&self) -> Sender<Point> {
+        self.sender.clone()
+    }
+    ///
+    /// Returns all `destinations` for the receiver
+    pub fn destinations(&self) -> Vec<PointDest> {
+        self.destinations.iter().map(|d| d.clone()).collect()
+    }
+    ///
+    /// Increments by one a count of the Receiver's subscriptions
+    pub fn inc(&self, destination: Option<impl Into<PointDest>>) {
+        self.subscriptions.fetch_add(1, Ordering::AcqRel);
+        if let Some(dest) = destination {
+            self.destinations.insert(dest.into());
+        }
+    }
+    ///
+    /// Decrements by one a count of the Receiver's subscriptions
+    pub fn dec(&self, destination: Option<impl Into<PointDest>>) {
+        if self.subscriptions.load(Ordering::Acquire) > 0 {
+            self.subscriptions.fetch_sub(1, Ordering::AcqRel);
+        }
+        if let Some(dest) = destination {
+            self.destinations.remove(&dest.into());
+        }
+    }
+    ///
+    /// Returns `true` if number of subscriptions is not a zero
+    pub fn is_subscribed(&self) -> bool {
+        self.subscriptions.load(Ordering::Acquire) > 0
     }
 }
